% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coleta_proposicoes_legislativa_ALMS.R
\name{coleta_proposicoes_legislativa_ALMS}
\alias{coleta_proposicoes_legislativa_ALMS}
\title{Coleta dados de proposições legislativa da Assembleia Legislativa
 do Mato Grosso do Sul, acessando a API do SGPL.}
\usage{
coleta_proposicoes_legislativa_ALMS(pagina = "1", n_proposicoes = "100")
}
\arguments{
\item{pagina}{argumento que controla a paginacao, o valor padrao é 1}

\item{n_proposicoes}{argumento que contorla o numero de
proposicoes por requisicao o padrao e 100}
}
\value{
um tibble com as proposicoes legislativas
}
\description{
Coleta dados de proposições legislativa da Assembleia Legislativa
 do Mato Grosso do Sul, acessando a API do SGPL.
}
\examples{
# Coleta as 100 primeiras proposicoes
# coleta_proposicoes_legislativa_ALMS(pagina = "1",
# n_proposicoes = "100")
}
